+++ src/Events.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * This file is part of the Hiberbee web application.
+ *
+ * (c) Vladyslav Volkov <vlad@hiberbee.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App;
+
+
+final class Events
+{
+    /**
+     * @Event("Symfony\Component\EventDispatcher\GenericEvent")
+     *
+     * @var string
+     */
+    const ACCOUNT_CREATED = 'account.created';
+}