+++ src/Kernel.php
@@ -34,7 +34,7 @@ class Kernel extends BaseKernel
      */
     public function getCacheDir()
     {
-        return $this->getProjectDir() . '/var/cache/' . $this->environment;
+        return $this->getProjectDir().'/var/cache/'.$this->environment;
     }
 
     /**
@@ -42,7 +42,7 @@ class Kernel extends BaseKernel
      */
     public function getLogDir()
     {
-        return $this->getProjectDir() . '/var/log';
+        return $this->getProjectDir().'/var/log';
     }
 
     /**
@@ -50,7 +50,7 @@ class Kernel extends BaseKernel
      */
     public function registerBundles()
     {
-        $contents = require $this->getProjectDir() . '/config/bundles.php';
+        $contents = require $this->getProjectDir().'/config/bundles.php';
         foreach ($contents as $class => $envs) {
             if (isset($envs['all']) || isset($envs[$this->environment])) {
                 yield new $class();
@@ -68,13 +68,13 @@ class Kernel extends BaseKernel
         $container->setParameter('app_locales', implode('|', LocaleEnumType::getChoices()));
         $container->setParameter('app_roles', implode('|', RoleEnumType::getChoices()));
 
-        $confDir = $this->getProjectDir() . '/config';
-        $loader->load($confDir . '/packages/*' . self::CONFIG_EXTS, 'glob');
-        if (is_dir($confDir . '/packages/' . $this->environment)) {
-            $loader->load($confDir . '/packages/' . $this->environment . '/**/*' . self::CONFIG_EXTS, 'glob');
+        $confDir = $this->getProjectDir().'/config';
+        $loader->load($confDir.'/packages/*'.self::CONFIG_EXTS, 'glob');
+        if (is_dir($confDir.'/packages/'.$this->environment)) {
+            $loader->load($confDir.'/packages/'.$this->environment.'/**/*'.self::CONFIG_EXTS, 'glob');
         }
-        $loader->load($confDir . '/services' . self::CONFIG_EXTS, 'glob');
-        $loader->load($confDir . '/services_' . $this->environment . self::CONFIG_EXTS, 'glob');
+        $loader->load($confDir.'/services'.self::CONFIG_EXTS, 'glob');
+        $loader->load($confDir.'/services_'.$this->environment.self::CONFIG_EXTS, 'glob');
     }
 
     /**
@@ -82,13 +82,19 @@ class Kernel extends BaseKernel
      */
     protected function configureRoutes(RouteCollectionBuilder $routes)
     {
-        $confDir = $this->getProjectDir() . '/config';
-        if (is_dir($confDir . DIRECTORY_SEPARATOR . self::ROUTE_CONFIG_DIRECTORY_NAME . DIRECTORY_SEPARATOR)) {
-            $routes->import($confDir . '/routes/*' . self::CONFIG_EXTS, '/', 'glob');
+        $confDir = $this->getProjectDir().'/config';
+        if (is_dir($confDir.DIRECTORY_SEPARATOR.self::ROUTE_CONFIG_DIRECTORY_NAME.DIRECTORY_SEPARATOR)) {
+            $routes->import($confDir.'/routes/*'.self::CONFIG_EXTS, '/', 'glob');
         }
-        if (is_dir($confDir . DIRECTORY_SEPARATOR . self::ROUTE_CONFIG_DIRECTORY_NAME . DIRECTORY_SEPARATOR . $this->environment)) {
-            $routes->import($confDir . DIRECTORY_SEPARATOR . self::ROUTE_CONFIG_DIRECTORY_NAME . DIRECTORY_SEPARATOR . $this->environment . '/**/*' . self::CONFIG_EXTS, '/', 'glob');
+        if (is_dir(
+            $confDir.DIRECTORY_SEPARATOR.self::ROUTE_CONFIG_DIRECTORY_NAME.DIRECTORY_SEPARATOR.$this->environment
+        )) {
+            $routes->import(
+                $confDir.DIRECTORY_SEPARATOR.self::ROUTE_CONFIG_DIRECTORY_NAME.DIRECTORY_SEPARATOR.$this->environment.'/**/*'.self::CONFIG_EXTS,
+                '/',
+                'glob'
+            );
         }
-        $routes->import($confDir . DIRECTORY_SEPARATOR . self::ROUTE_CONFIG_DIRECTORY_NAME . self::CONFIG_EXTS, '/', 'glob');
+        $routes->import($confDir.DIRECTORY_SEPARATOR.self::ROUTE_CONFIG_DIRECTORY_NAME.self::CONFIG_EXTS, '/', 'glob');
     }
 }